Harden Vulkan lifecycle, culling, frame uploads, and smoke tests
- Pin Vulkan pipeline creation arrays and make renderer disposal resilient
- Replace FNV geometry hashing with per-frame byte snapshots
- Add viewport culling and minimized-window render skipping
- Improve DPI awareness, entity generations, and cap-flag parsing
- Expand smoke tests and update documentation/context

+++ src/Engine.Core/EntityId.cs
@@ -1,6 +1,6 @@
 namespace Engine.Core;
 
-public readonly record struct EntityId(uint Index, ushort Generation)
+public readonly record struct EntityId(uint Index, uint Generation)
 {
     public static readonly EntityId Invalid = new(uint.MaxValue, 0);
     public bool IsValid => Index != uint.MaxValue;
